Moving StorageConfiguration.java out of the internal package
