+++ disk-buffering/src/main/java/io/opentelemetry/contrib/disk/buffering/StorageConfiguration.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.contrib.disk.buffering.internal;
+package io.opentelemetry.contrib.disk.buffering;
 
 import com.google.auto.value.AutoValue;
 import io.opentelemetry.contrib.disk.buffering.internal.files.DefaultTemporaryFileProvider;
